Extracted dispatching of actions to separate class.
